refactor: autofocus the first cell of the new row
