--- packages/rich-text/src/plugins/Table/actions/addRowAbove.ts
@@ -1,37 +1,0 @@
-import { getAbove, insertNodes, someNode } from '@udecode/slate-plugins-common';
-import { getSlatePluginType, SPEditor, TElement } from '@udecode/slate-plugins-core';
-import {
-  ELEMENT_TABLE,
-  ELEMENT_TR,
-  TablePluginOptions,
-  getEmptyRowNode,
-} from '@udecode/slate-plugins-table';
-
-export const addRowAbove = (editor: SPEditor, { header }: TablePluginOptions) => {
-  if (
-    someNode(editor, {
-      match: { type: getSlatePluginType(editor, ELEMENT_TABLE) },
-    })
-  ) {
-    const currentRowItem = getAbove(editor, {
-      match: { type: getSlatePluginType(editor, ELEMENT_TR) },
-    });
-
-    if (currentRowItem) {
-      const [currentRowElem, currentRowPath] = currentRowItem;
-
-      insertNodes<TElement>(
-        editor,
-        // @ts-expect-error
-        getEmptyRowNode(editor, {
-          header,
-          colCount: currentRowElem.children.length,
-        }),
-        {
-          at: currentRowPath,
-          select: true,
-        }
-      );
-    }
-  }
-};
